+++ internal/models/permission_groups.go
@@ -3,7 +3,7 @@ package models
 import "time"
 
 type PermissionGroup struct {
-	Id        int       `db:"id" json:"id"`
+	Id        int64     `db:"id" json:"id"`
 	Name      string    `db:"name"  json:"name"`
 	CreatedAt time.Time `db:"created_at"  json:"created_at"`
 	UpdatedAt time.Time `db:"updated_at" json:"updated_at"`
